LayoutTests/media/media-source/media-source-webm-fragment.html

<!DOCTYPE html> <!-- webkit-test-runner [ MediaSourceEnabled=true ] -->
<html>
<head>
    <title>media-source-webm-fragment</title>
    <script src="../../media/media-source/media-source-loader.js"></script>
    <script src="../../media/video-test.js"></script>
    <script>
    var loader;
    var source;
    var sourceBuffer;
    var mediaSegmentDuration;

    function loaderPromise(loader) {
        return new Promise((resolve, reject) => {
            loader.onload = resolve;
            loader.onerror = reject;
        });
    }

    window.addEventListener('load', async event => {
        try {
            findMediaElement();

            let manifests = [ 'content/test-opus-manifest.json', 'content/test-vorbis-manifest.json' ];
            for (const manifest of manifests) {
                loader = new MediaSourceLoader(manifest);
                await loaderPromise(loader);
                if (MediaSource.isTypeSupported(loader.type()))
                    break;
            }
            mediaSegmentDuration = loader.mediaSegmentEndTime(0);
            source = new MediaSource();
            run('video.src = URL.createObjectURL(source)');
            await waitFor(source, 'sourceopen');
            waitFor(video, 'error').then(failTest);

            run('sourceBuffer = source.addSourceBuffer(loader.type())');

            run('sourceBuffer.appendBuffer(loader.initSegment())');
            await waitFor(sourceBuffer, 'update');

            for (let i = 1; i <= 10; i++) {
                run('sourceBuffer.appendBuffer(loader.mediaSegment(0))');
                await waitFor(sourceBuffer, 'update');
                run('sourceBuffer.timestampOffset = sourceBuffer.buffered.end(sourceBuffer.buffered.length - 1)');
            }
            run('source.endOfStream()');
            await waitFor(source, 'sourceended');
            // Compare against 9* the size, to account for overlaps that could happen depending on implementation.
            test('source.duration >= mediaSegmentDuration*9');
            endTest();
        } catch (e) {
            failTest(`Caught exception: "${e}"`);
        }
    });
    </script>
</head>
<body>
    <video controls></video>
</body>
</html>
